// src/app/types/tour.ts
// /src/app/types/tour.ts
export type Tour = {
    id: number
    name: string
    image: string
    description: string
    duration: string
    destination: string
    collection: string | null
    price: number
    link: string
    travelForCredit: boolean
    aPlusCollection: boolean
    popularity: number;
  }
  